LayoutTests/http/tests/resourceLoadStatistics/prevalent-resource-handled-keydown-database-expected.txt

Tests that we grant User Interaction credit for handled keypresses.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Simulate user typing letter 'a' into the field.
PASS testInput.value is "a"
PASS Origin was granted user interaction.
PASS successfullyParsed is true

TEST COMPLETE
